delete extensions.py module
